Move the arenas state to the runtime state.
